Split manifest internals behind the existing public facade
